Add support for template-filing of bugs.
What this allows is to have a quick way to file bugs by creating a [B]
link that pre-fills most of the bug data for you, including the
assignees, the product, the URL to the log, and so on.

The main comment is set to the `emerge --info` output that has to be
present in the emerge-infos directory with the reverse resolution of
the hostname reporting the log.